fix: align and distribute binded text in container and cleanup
